backend/requirements.txt
weaviate-client
llama-index
python-dotenv
openai
black
wasabi
typer
haystack-ai
nltk
fastapi
uvicorn